move misplaced tests
